refactor(cli): move the wizard's everos cluster into its own module
onboard_commands.py had grown past 5000 lines; the EverOS/Step4 cluster
(~1500 lines, 36 symbols) now lives in onboard_everos.py. Pure move,
zero behavior change: shared wizard UI state stays in onboard_commands
and both modules reach across through module references with call-time
attribute access, so import order cannot bite and monkeypatches keep
hitting the code that actually runs. Every migrated patch target in the
test file was mutation-checked one by one against pointing at the old
module -- all 23 turn red, none patch a shell.
